Added answer for using trigger to stop PA system.
Not the best way of doing it. In a normal implementation it would be better to have a dedicated actor for the control panel. It would also be better to not rely on the level blueprint for spawning sounds at given locations. An actor should be able to take care of this for the sake of reusability.